reorganized files
- moved `maze` folder into a `general` folder
- moved `helpers` folder into `tests` directory
