+++ tests/general/maze/test_cell.py
